requirements.txt
fastapi
uvicorn[standard]
sse-starlette
pydantic-ai
gitpython
google-generativeai
anthropic
openai
modal
PyGithub
python-dotenv
requests
